Took LDAP functionality into base class for DUnit
