+++ cluster/src/dunit/scala/io/snappydata/cluster/ClusterManagerLDAPTestBase.scala
@@ -1,0 +1,63 @@
+/*
+ * Copyright (c) 2016 SnappyData, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you
+ * may not use this file except in compliance with the License. You
+ * may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * permissions and limitations under the License. See accompanying
+ * LICENSE file.
+ */
+package io.snappydata.cluster
+
+import java.util.Properties
+
+import scala.language.postfixOps
+
+import com.pivotal.gemfirexd.Attribute
+import com.pivotal.gemfirexd.security.{LdapTestServer, SecurityTestUtils}
+
+/**
+ * Base class for start and stop of LDAP Server
+ *
+ * @author vivek
+ */
+
+abstract class ClusterManagerLDAPTestBase(s: String, adminUser: String)
+    extends ClusterManagerTestBase(s) with Serializable {
+
+  override def beforeClass(): Unit = {
+    val ldapProperties = SecurityTestUtils.startLdapServerAndGetBootProperties(0, 0, adminUser,
+      getClass.getResource("/auth.ldif").getPath)
+    setSecurityProps(ldapProperties)
+    super.beforeClass()
+  }
+
+  override def afterClass(): Unit = {
+    val ldapServer = LdapTestServer.getInstance()
+    if (ldapServer.isServerStarted) {
+      ldapServer.stopService()
+    }
+    super.afterClass()
+  }
+
+  def setSecurityProps(ldapProperties: Properties): Unit = {
+    import com.pivotal.gemfirexd.Property.{AUTH_LDAP_SERVER, AUTH_LDAP_SEARCH_BASE}
+    for (k <- List(Attribute.AUTH_PROVIDER, AUTH_LDAP_SERVER, AUTH_LDAP_SEARCH_BASE)) {
+      System.setProperty(k, ldapProperties.getProperty(k))
+    }
+    for (k <- List(Attribute.AUTH_PROVIDER, AUTH_LDAP_SERVER, AUTH_LDAP_SEARCH_BASE,
+      Attribute.USERNAME_ATTR, Attribute.PASSWORD_ATTR)) {
+      locatorNetProps.setProperty(k, ldapProperties.getProperty(k))
+      bootProps.setProperty(k, ldapProperties.getProperty(k))
+    }
+  }
+}
+
+
